requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
kafka-python==1.4.6
numpy==1.16.4
requests==2.22.0
urllib3==1.25.3
